Initial refactor
Added new packages and moved some files around because Git wouldn't recognize the new directories.
